Select tour origin and destination from type groups
Origin and destination selection is changed from HEHA location types to
type groups. The number of the group is also the priority for choosing
origins and destinations. Origin may be different from the source of
the tour which, in turn, is always either the start and return point of
a closed tour, or home if the tour is open, or missing, if home is not
visited.