Simplify tour type definition
The first definition of tour type was too complicated so we will try if
this one is better.

First, take the type groups of the locations that are visited during
the tour. (If a person visits two shops, then its type group is in the
list twice.) Then, sort those groups by their priorities, and there you
have tour types.