Change survey-specific person ID columns to common column
We wish to ensure that two people from different surveys can not have
identical person ids. That is why we changed input scripts to write new
ids for every person. This column 'pid' contains that new id. The actual
script is an R file which will be uploaded to Github once the strafica
package is made public.