Fix destination search when coordinates are missing
Finding destinations worked incorrectly if there were locations with
missing (nan) coordinates. This bug was found when developing a data set
with secondary destinations. Example: a person with pid=100672 made one
trip on survey date which was from leisure (9) to home (1). This
open-ended tour should have home as origin and leisure as destination
and no secondary destination. Instead, the tour had home as both origin
and destination, and leisure as secondary destination. This commit fixes
this bad behavior!