Generate alternatives from secondary destinations
Written separately because the calculation of travel times and cost
differ from other alternatives. Here, travel times and costs for a
tour A-B-C (A origin, B destination, C secondary destination) are
defined as BC + CA - BA which gives us the extra time/cost of visiting
a second destination instead of just one.