Fix condition for other destinations in a tour
As we are handling tours with origin, primary destination, and secondary destination (= 3 places), a tour will have other destinations if a person visits more than three places. during a tour.As we are handling tours with origin, primary destination, and secondary
destination (= 3 places), a tour will have other destinations if a person
visits more than three places. during a tour.